extract block to mod
